+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2020-06-13
+### Added
+- New Order types support : Trailing stop, Trailing stop limit, Take profit, Take profit limit
+### Updated
+- Order fill and post fill refactor
+### Fixed
+- Tests silent exceptions
+
 ## [1.7.2] - 2020-06-01
 ### Added
 - PriceEventManager